rename: components renaming
