+++ packages/web-domains/src/common/components/KakaoShare/CopyLink.styles.ts
