+++ packages/web-domains/src/common/components/KakaoShare/ToastCheckIcon.tsx
@@ -1,13 +1,8 @@
-import { Interpolation, Theme } from '@emotion/react';
 import { HTMLAttributes } from 'react';
 
-interface ToastCheckProps extends HTMLAttributes<SVGSVGElement> {
-  css?: Interpolation<Theme>;
-}
-
-export const ToastCheck = ({ css, ...rest }: ToastCheckProps) => {
+export const ToastCheckIcon = (rest: HTMLAttributes<SVGSVGElement>) => {
   return (
-    <svg width="40" height="41" viewBox="0 0 40 41" fill="none" xmlns="http://www.w3.org/2000/svg" css={css} {...rest}>
+    <svg width="40" height="41" viewBox="0 0 40 41" fill="none" xmlns="http://www.w3.org/2000/svg" {...rest}>
       <g clipPath="url(#clip0_3987_16110)">
         <path
           d="M37.199 7.88501L14.1674 30.915C14.0125 31.0705 13.8285 31.1938 13.6258 31.278C13.4232 31.3622 13.2059 31.4055 12.9865 31.4055C12.7671 31.4055 12.5498 31.3622 12.3472 31.278C12.1446 31.1938 11.9605 31.0705 11.8057 30.915L2.89902 22C2.74417 21.8446 2.56015 21.7212 2.35751 21.637C2.15488 21.5529 1.93761 21.5095 1.71819 21.5095C1.49877 21.5095 1.2815 21.5529 1.07887 21.637C0.876229 21.7212 0.692206 21.8446 0.537357 22V22C0.381896 22.1549 0.258542 22.3389 0.174372 22.5415C0.0902028 22.7442 0.046875 22.9614 0.046875 23.1808C0.046875 23.4003 0.0902028 23.6175 0.174372 23.8202C0.258542 24.0228 0.381896 24.2068 0.537357 24.3617L9.44736 33.27C10.3873 34.2082 11.661 34.7351 12.989 34.7351C14.317 34.7351 15.5908 34.2082 16.5307 33.27L39.5607 10.245C39.7159 10.0902 39.839 9.90628 39.9231 9.70379C40.0071 9.50131 40.0503 9.28424 40.0503 9.06501C40.0503 8.84579 40.0071 8.62872 39.9231 8.42623C39.839 8.22375 39.7159 8.03983 39.5607 7.88501C39.4058 7.72955 39.2218 7.6062 39.0192 7.52203C38.8165 7.43786 38.5993 7.39453 38.3799 7.39453C38.1604 7.39453 37.9432 7.43786 37.7405 7.52203C37.5379 7.6062 37.3539 7.72955 37.199 7.88501Z"
